Fix compiler warnings/errors with modern compilers
Newer versions of gcc/clang were throwing warnings in a few piaces,
this patch does the minimum necessary to make the compiler happy
by adding appropriate casts in line with what was present previous.

The sscanf bits, in particular, do appear to have some fundamental
safety issues and checking that are absent and might be worth a second
look as with malicious input the results could be unexpected.

Signed-off-by: Paul Osborne <osbpau@gmail.com>